Major bugfix!  Was getting different hashes foor hash tables greater than 65K in length (such as in the Russian dicts!) Phew. This one was hard to find.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@33099 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ link-grammar/structures.h
@@ -162,7 +162,7 @@ typedef enum
 
 struct Connector_struct
 {
-    short hash;
+    int hash;
     unsigned char word;
                    /* The nearest word to my left (or right) that
                       this could connect to.  Computed by power pruning */
